+++ src/Settings/PreferencesWindow.cpp
@@ -116,6 +116,13 @@ PreferencesWindow::PreferencesWindow(QWidget *parent) : QMainWindow(parent)
 
     addPage("Extensions/Clang Format", {"Clang Format/Path", "Clang Format/Style"}, false);
 
+    addPage("Extensions/Language Server/C++ Server",
+            {/*"LSP/Use Autocomplete C++",*/ "LSP/Use Linting C++", "LSP/Delay C++", "LSP/Path C++", "LSP/Args C++"});
+    addPage("Extensions/Language Server/Java Server", {/*"LSP/Use Autocomplete Java",*/ "LSP/Use Linting Java",
+                                                       "LSP/Delay Java", "LSP/Path Java", "LSP/Args Java"});
+    addPage("Extensions/Language Server/Python Server", {/*"LSP/Use Autocomplete Python",*/ "LSP/Use Linting Python",
+                                                         "LSP/Delay Python", "LSP/Path Python", "LSP/Args Python"});
+
     addPage("Extensions/Competitive Companion", {"Competitive Companion/Enable", "Competitive Companion/Open New Tab",
                                                  "Competitive Companion/Connection Port"});