+++ src/Util.hpp
@@ -57,6 +57,16 @@ void setEditorLanguage(QCodeEditor *editor, const QString &language);
 
 QPalette windowsDarkThemePalette();
 
+enum ArgumentSplitterStatus
+{
+    SA_OUT,
+    SA_NORMAL,
+    SA_SQUOTE, // single quote
+    SA_DQUOTE  // double quote
+};
+
+QStringList splitArgument(QString);
+
 } // namespace Util
 
 #endif // UTIL_HPP