+++ doc/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## UNRELEASED
 
+### Linting with Language Server
+
+Now error linting is available with Language Server. Linting can be helpful when you want to find errors in your code and fix them quickly. You need to install a Language Server program to enable linting. You can set delay for linting, the path to the Languae Server program and its arguments in the Preferences window. You can set different settings for different languages. Please check the [manual](MANUAL.md) for more information on how to set this up.
+
 ### Added
 
 - Add Ctrl+Shift+D for duplicating the current line or the selection. (#237)
@@ -10,6 +14,7 @@
 
 ### Changed
 
+- Word selection in editor is highlighted with change of background color instead of an underline. Underlined text now represent linting of type "Hint".
 - The shortcut of Detached Run is changed from Ctrl+Shift+D to Ctrl+Alt+D. (#237)
 
 ### Fixed